minor changes & refactor
