+++ .gitignore
@@ -2,3 +2,5 @@
 .$Hero*
 .vscode
 __pycache__
+venv
+.venv
