+++ data/data.py
@@ -1,0 +1,32 @@
+weaknesses = [
+    "slash",
+    "blunt",
+    "stab",
+    "fire",
+    "water",
+    "wind",
+    "ice",
+    "poison",
+    "electric",
+]
+resistances = [
+    "slash",
+    "blunt",
+    "stab",
+    "fire",
+    "water",
+    "wind",
+    "ice",
+    "poison",
+    "electric",
+]
+
+states = [
+    "frozen",
+    "burn",
+    "poisoned",
+    "shocked",
+    "scared",
+    "sleep",
+    "blind",
+]
