+++ data/weapons.py
